+++ CHANGELOG.md
@@ -7,17 +7,49 @@ All notable changes to this project will be documented in this file. This projec
 
 ### Added
 
-- **BREAKING** The `Contracts\Query\QueryParameters` interface now has an `unrecognisedParameters` method. This returns
-  any query parameters that are not defined by the JSON:API specification, which allows implementations to add support
-  for additional query parameters as needed.
+- **BREAKING** The `Contracts\Schema\Schema` interface now has `isFilter()`, `isSparseField` and `isSortable()` methods.
+  These methods have been added to the abstract schema class provided by this package, so this is unlikely to have a
+  significant impact on implementing packages.
+- **BREAKING** Made the following changes to the `Contracts\Query\QueryParameters` interface:
+    - New `unrecognisedParameters` method. This returns any query parameters that are not defined by the JSON:API
+      specification, which allows implementations to add support for additional query parameters as needed.
+    - The `filters` method now returns a `FilterParameters` object or null. Previously it returned an array or null.
+- New `FilterParameters` class for handling a collection of filter parameters received from a client.
 - The `FieldSets`, `IncludePaths` and `SortFields` classes all now have a `collect()` method, that returns a collection
   object.
-- The `IncludePaths` class now has `filter` and `reject` methods.
+- The `IncludePaths` and `SortFields` classes now have `filter`, `reject` and `forSchema` methods.
 - The `SortField` class now has static `ascending` and `descending` methods, to easily create a sort field with the
   specified direction.
 - The `QueryParameters` class now has a `toQuery()` method, that casts the value back to a HTTP query parameter array.
   This is different from `QueryParameters::toArray()`, as the `include` and `sort` parameters are strings in a HTTP
   query array.
+- The `QueryParameters` class now has a `forSchema()` method, that returns a new query parameters instance that contains
+  only parameters valid for the supplied schema.
+- The `Document\ResourceObject` class has a new `withRelationshipMeta` method for adding meta for a specified
+  relationship.
+- Added new response classes for returning related resources for a relationship - e.g. the `/api/posts/1/comments`
+  endpoint. Previously the `DataResponse` class was used for this endpoint, but the new classes allow for relationship
+  meta to be merged into the top-level meta member of the response for the endpoint.
+- The core package now supports the *countable* implementation-semantic. This adds a custom query parameter that allows
+  a client to specify which relationships should have a count added to their relationship meta.
+
+### Changed
+
+- **BREAKING** The `Contracts\Encoder\Encoder` interface now has two methods for encoding resource identifiers:
+  `withToOne` and `withToMany`. These replace the `withIdentifiers` method, which has been removed.
+- Moved the following classes from the `Core\Responses` namespace to the `Core\Responses\Internal` namespace. This is
+  considered non-breaking because the classes are not part of the public API (responses that can be used for the public
+  API are still in the `Core\Responses` namespace):
+    - `PaginatedResourceResponse`
+    - `ResourceCollectionResponse`
+    - `ResourceIdentifierCollectionResponse`
+    - `ResourceIdentifierResponse`
+    - `ResourceResponse`
+
+### Removed
+
+- Deleted the `Core\Responses\Concerns\EncodesIdentifiers` trait. This is considered non-breaking as the trait was only
+  intended for internal use.
 
 ### Fixed